requirement.txt
torch
torchvision
torchaudio
ultralytics
transformers
opencv-python
pillow
pyyaml
matplotlib
tqdm